Create Mediodía
Dark Waters is a series of drawings that attempt to represent states of my mind during meditation processes. They manifest energetic shocks, spatial translations, explosions of color and packets of information that happen in my psyche during transcendent processes.

Noon reproduces the chaos of information and emotions that occur in the waking state

The code used is a fork of Landscape by Henrik Linusson, included in his manual of examples for Processing. You can check it here: