Sunset
Dark Waters is a series of drawings that attempt to represent states of my mind during meditation processes. They manifest energetic shocks, spatial translations, explosions of color and packets of information that happen in my psyche during transcendent processes.

Sunset shows the state of balance between the contents of consciousness and the stillness of meditation; it is a state that includes everyday reality but shows it for what it is: a vibratory flow within the being.